# src/main/java/it/muretti/micro/controller/WriteController.java
package it.muretti.micro.controller;

import java.time.ZonedDateTime;
import java.time.format.DateTimeFormatter;
import java.util.Date;


import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;


import dto.AddRapperRequest;

import it.muretti.micro.entity.MurettiFreestyleEntity;
import it.muretti.micro.entity.Presenza;

import it.muretti.micro.request.RequestPresenza;
import it.muretti.micro.service.MurettiFreestyleService;
import org.springframework.web.bind.annotation.*;

@CrossOrigin(origins = "http://localhost:3000") // Permette a tutto il controller di accettare richieste da React
@RestController
@RequestMapping("/murettifreestyle")
public class WriteController {
	
	
	
	@Autowired
	private MurettiFreestyleService murettifreestyleService;
	
	@PostMapping("/add")
    public ResponseEntity<MurettiFreestyleEntity> addUtente(@RequestBody MurettiFreestyleEntity entity) {
		try {
	        MurettiFreestyleEntity nuovoMuretto = murettifreestyleService.addEntity(entity);
	        // Restituisci la risposta con codice 201 per successo
	        return ResponseEntity.status(HttpStatus.CREATED).body(nuovoMuretto);
	    } catch (Exception e) {
	        // Gestisci errori se ci sono
	        e.printStackTrace();
	        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
	    }
    }
	
	 // aEndpoint per aggiungere la presenza a un rapper specifico
     @PostMapping("/addPresenza/{tipo}/{valore}/{rapperNome}")
     public ResponseEntity<?> addPresenza(
             @PathVariable String tipo,
             @PathVariable String valore,
             @PathVariable String rapperNome,
             @RequestBody RequestPresenza presenza) {



         try {
             boolean added = murettifreestyleService.addPresenzaToRapper(tipo, valore, rapperNome, presenza);
             if (added) {
                 System.out.println("Dati ricevuti: " + presenza.toString());
                 return ResponseEntity.ok("Presenza aggiunta correttamente");
             } else {
                 return ResponseEntity.status(404).body("Rapper non trovato o errore nel processo");
             }
         } catch (IllegalArgumentException e) {
             return ResponseEntity.badRequest().body("Errore: " + e.getMessage());
         }
     }


    @PutMapping("/updatePresenza")
    public ResponseEntity<?> updatePresenza(
        @RequestParam String tipo, // muretto
        @RequestParam String valore, //messina
        @RequestParam String nome,  //resla
        @RequestParam String data,  // data
        @RequestBody Presenza nuovaPresenza) {

        DateTimeFormatter formatter = DateTimeFormatter.ISO_OFFSET_DATE_TIME;
        data = data.replace(" ", "+");

        Date dataConvertita=null;
        try {
            ZonedDateTime dataPresenza = ZonedDateTime.parse(data, formatter);
            dataConvertita = Date.from(dataPresenza.toInstant());

        } catch (Exception e) {
            e.printStackTrace();
            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Formato data non valido: " + data);
        }

        // Passaggio dei dati aggiornati
        boolean updated = murettifreestyleService.updatePresenza(tipo, valore,nome, dataConvertita, nuovaPresenza ,  nuovaPresenza.getEvento(), nuovaPresenza.getPunteggio());
        if (updated) {
        	System.out.println(nuovaPresenza.toString());
            return ResponseEntity.ok("Presenza aggiornata correttamente");


        } else {
            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Presenza non trovata o errore nel processo");
        }


    }

    @DeleteMapping("/{valore}/{nome}")
    public ResponseEntity<?> deletePresenza(
        
        @PathVariable String valore,
        @PathVariable String nome,
        @RequestParam String data
        ) {
    	
    	DateTimeFormatter formatter = DateTimeFormatter.ISO_OFFSET_DATE_TIME;
        data = data.replace(" ", "+");
        
        Date dataConvertita=null;
        try {
            ZonedDateTime dataPresenza = ZonedDateTime.parse(data, formatter);
            dataConvertita = Date.from(dataPresenza.toInstant());
            
        } catch (Exception e) {
            e.printStackTrace();
            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Formato data non valido: " + data);
        }

    	  boolean deleted = murettifreestyleService.deletePresenza(valore, nome, dataConvertita);
          if (deleted) {
              return ResponseEntity.ok("Data eliminata con successo!");
          } else {
              return ResponseEntity.notFound().build();
          }    
    }
    
    @PostMapping("/addRapper")
    public ResponseEntity<String> addRapper(@RequestBody AddRapperRequest request) {
        boolean success = murettifreestyleService.newRapperToMuretto(
                request.getValore(),
                request.getAlias(),
                request.getRapper()
        );

        if (success) {
            return ResponseEntity.ok("Rapper aggiunto con successo!");
        } else {
            return ResponseEntity.badRequest().body("Errore: Nessun documento trovato o aggiornato.");
        }
    }
    @DeleteMapping("/deleteRapper/{valore}/{alias}")
    public ResponseEntity<?> deleteRapper(
        
        @PathVariable String valore,
        @PathVariable String alias,
        @RequestParam String nome
        
        ) {
    	
    	  boolean deleted = murettifreestyleService.deleteRapper(valore, nome, alias);
          if (deleted) {
              return ResponseEntity.ok("Rapper eliminato con successo!");
          } else {
              return ResponseEntity.notFound().build();
          }    
    }
    
    @PutMapping("/updateRapper/{tipo}/{valore}/{oldName}")
    public ResponseEntity<?> updateRap(
    	@PathVariable String tipo,
    	@PathVariable String valore,
        @PathVariable String oldName,
        @RequestParam String nome,
        @RequestParam int rank
          
        ) {

       
        // Passaggio dei dati aggiornati
        boolean updated = murettifreestyleService.updateRapper(tipo, valore, oldName,  nome , rank);
        if (updated) {
        	
            return ResponseEntity.ok("Rapper aggiornato correttamente");
           
            
        } else {
            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Presenza non trovata o errore nel processo");
        }
        
        
    }


}